Integration tests for STOMP 1.2 pass for all current versions of supported brokers:
* ActiveMQ 5.8
* RabbitMQ 3.0.2
* Apollo 1.5 (still doesn't properly support heart-beating)

Added some new test cases and fixed a couple of existing ones.
Added one bugfix to catch a protocl error when the async client
optimistically NACKs a frame even when the session's version is 1.0
(this way, the client doesn't have to know which version its session is
in which is a good thing).