async: store ERROR frame when disconnecting
When the async client sees an ERROR frame, we immediately disconnect()
with a StompProtocolError reason. This StompProtocolError contains a
textual description of the frame (frame.info()) so errbacks can log
something, but it is not possible for errbacks to get specific data
from the problematic ERROR frame.

Store the frame as an attribute of our StompProtocolError when we
disconnect(). This allows errbacks to inspect the bad frame.

The practical use-case here is that ActiveMQ will return an ERROR frame
if a user is not permitted to access a certain topic, and it's nice to
inspect the ERROR frame's specific "message" header in an errback to
print the exact ActiveMQ error message to the user.